requirements.txt
pyautogui
Pillow
easyocr
opencv-python
torch
torchvision
torchaudio
